feat: add SideBar component
