+++ src/components/SideBar/SideBar.module.scss
@@ -1,0 +1,17 @@
+.side-bar {
+  height: 100%;
+  margin: 0.5rem 0 1rem 1rem;
+
+  .side-bar-controls {
+    height: 100%;
+    position: relative;
+    z-index: 10;
+  }
+
+  .side-bar-viewer {
+    height: 100%;
+    overflow: scroll;
+    position: relative;
+    z-index: 0;
+  }
+}
