+++ src/components/SideBar/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SideBar';
