+++ src/pages/MainPage/MainPage.tsx
@@ -1,40 +1,15 @@
-import React, { useState, useContext } from "react";
-import { Button, Layout } from "antd";
-import { MenuFoldOutlined, MenuUnfoldOutlined } from "@ant-design/icons";
-import styles from "./MainPage.module.scss";
+import React from "react";
+import { Layout } from "antd";
 import CodePlayground from "../../components/CodePlayground";
-import Sidebar from "../../components/Sidebar";
-import { StoreContext } from "../../store/StoreProvider";
+import SideBar from "../../components/SideBar";
+import styles from "./MainPage.module.scss";
 
-const { Sider, Content } = Layout;
+const { Content } = Layout;
 
 const MainPage: React.FC = () => {
-  const [collapsed, setCollapsed] = useState(true);
-  const store = useContext(StoreContext);
-
-  const handleRequest = () => {
-    store.schemaStore.sendRequest();
-  };
-
   return (
     <Layout className={styles["main-page"]}>
-      <Sider
-        trigger={null}
-        theme="light"
-        collapsible
-        collapsed={collapsed}
-        width={500}
-      >
-        <Button
-          onClick={() => {
-            setCollapsed(!collapsed);
-            collapsed && handleRequest();
-          }}
-        >
-          {collapsed ? <MenuUnfoldOutlined /> : <MenuFoldOutlined />}
-        </Button>
-        {!collapsed && <Sidebar />}
-      </Sider>
+      <SideBar />
       <Content className={styles["page-content"]}>
         <CodePlayground />
       </Content>
